feat : Blog 컴포넌트 생성 후 적용
