+++ packages/blog/src/pages/Main/Blog/Blog.module.scss
@@ -1,0 +1,19 @@
+.wrapper {
+  width: 100%;
+  margin-bottom: 20px;
+
+  .blogTitle {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+    align-items: baseline;
+  }
+
+  .boxWrapper {
+    width: 100%;
+    display: flex;
+    justify-content: baseline;
+    gap: 10px;
+    flex-wrap: wrap;
+  }
+}
