+++ packages/blog/src/pages/Main/Blog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Blog';
