--- packages/blog/src/pages/Main/Main.module.scss
@@ -1,18 +1,0 @@
-.header {
-  justify-content: center;
-}
-
-.boxWrapper {
-  width: 100%;
-  display: flex;
-  justify-content: baseline;
-  gap: 10px;
-  flex-wrap: wrap;
-}
-
-.blogTitle {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: baseline;
-}
